Silencer now uses one file, like the rest of the heroes
